requirements.txt
mysql-connector-python
dnspython
requests
